Profile funtionality complete
